backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "RAG demo backend with FastAPI and LangGraph"
authors = [
    { name = "Michael Hua", email = "michael.huayl@yahoo.com" }
]
readme = "README.md"
package-mode = false
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "nanoid (>=2.0.0,<3.0.0)",
    "langchain (>=0.3.23,<0.4.0)",
    "langgraph (>=0.3.31,<0.4.0)",
    "langchain-community (>=0.3.21,<0.4.0)",
    "pymupdf (>=1.25.5,<2.0.0)",
    "python-pptx (>=1.0.2,<2.0.0)",
    "tiktoken (>=0.9.0,<0.10.0)",
    "langchain-weaviate (>=0.0.4,<0.0.5)",
    "langchain-google-genai (>=2.1.3,<3.0.0)",
    "ipython (>=9.2.0,<10.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
ruff = "^0.11.6"
pytest-mock = "^3.14.0"
pytest-asyncio = "^0.26.0"

[tool.ruff]
show-fixes = true
target-version = "py312"
line-length = 80

[tool.ruff.lint]
extend-select = ["N", "UP", "PL", "Q"]
ignore = ["E501"]

